# api/openapi.yaml
openapi: 3.1.0
info:
  title: AppHub control plane
  version: 1.0.0
  description: |
    Organization application and durable deployment API. This document is the wire
    contract for browser, CLI, SDK, and MCP adapter inputs and results. All API
    paths are relative to the operator-configured public origin. Production uses
    HTTPS; HTTP is allowed only on an explicitly configured literal loopback
    development origin. Origins and OAuth resources are never inferred from Host
    or forwarded headers.

    Users authenticate through an admitted Google or OIDC identity. Exact issuer
    and subject identify an account; equal email addresses never merge accounts.
    Eligibility, local disablement, admission, and operator-configured admin role
    are checked on each request. Ordinary users can access only their own
    applications and deployments. Admins can operate any application. Inaccessible
    object IDs return 404, before revealing scope failures on another owner's
    objects. Known accessible objects with insufficient scope return 403.

    API requests support a browser session OR an opaque AppHub OAuth bearer token
    for the exact resource {publicOrigin}/api. POST /api/v1/github/webhook is the
    exception: GitHub authenticates with X-Hub-Signature-256, and a session or
    bearer token is not a credential there. If Authorization is present, bearer
    authentication is selected exclusively; an invalid bearer never falls back to
...

...

...

...
servers:
  - url: /
...
security:
  - SessionCookie: []
  - OAuth: []
paths:
...
  /.well-known/oauth-protected-resource/mcp:
    get:
      operationId: getMCPResourceMetadata
      tags: [OAuth]
      summary: Discover the exact remote MCP OAuth resource
      security: []
      responses:
        '200':
          description: resource is exactly {publicOrigin}/mcp; an API token is not valid here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtectedResourceMetadata'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-apphub_session
      description: |
        Host-only Secure HttpOnly SameSite=Lax session, Path=/, no Domain. Explicit
        loopback HTTP uses apphub_session instead. Cookie mutations require exact
        Origin and X-CSRF-Token. Never accepted as an OAuth token or at /mcp.
    OAuth:
      type: oauth2
      description: |
        Opaque AppHub bearer tokens only. API operations require resource exactly
        {publicOrigin}/api; MCP requires {publicOrigin}/mcp. Send the resource in
        both authorization and code exchange. Scope names are exact set members,
        not prefixes. Every token must have at least one known scope. The empty
        requirement list on user/session operations means any nonempty granted
        scope is sufficient, not an unscoped token. Browser sessions do not need
        OAuth scopes. Invalid Authorization never falls back to a cookie.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            applications:read: Read accessible applications and safe target options.
            applications:write: Create and edit accessible applications.
            deployments:read: Read accessible deployment operations and history.
            deployments:write: Submit deployments; resolve interrupted work only if also admin.
...
  responses:
...
    Unavailable:
      description: Required storage, authentication provider, target configuration, or worker readiness unavailable; retryable, never an empty success.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable service-authored classification, never a raw provider error.
        message:
          type: string
          description: Safe user-facing explanation, not a stack trace or build log.
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          description: Editable field path to safe validation explanation.
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: '#/components/schemas/Error'
        requestId:
          type: string
...
    Scope:
      type: string
      enum: ['applications:read', 'applications:write', 'deployments:read', 'deployments:write']
...
    OAuthResource:
      type: string
      format: uri
      description: Exactly {publicOrigin}/api, {publicOrigin}/mcp, or an enabled hosted application's https://{applicationHost}/mcp audience; one resource per token/family, never a prefix.
...
    ProtectedResourceMetadata:
      type: object
      required: [resource, authorization_servers, scopes_supported, bearer_methods_supported]
      properties:
        resource:
          $ref: '#/components/schemas/OAuthResource'
        authorization_servers:
          type: array
          minItems: 1
          maxItems: 1
          items:
            type: string
            format: uri
          description: Exactly one entry, the configured publicOrigin issuer.
        scopes_supported:
          type: array
          minItems: 4
          maxItems: 4
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Scope'
        bearer_methods_supported:
          type: array
          items:
            type: string
            const: header
          const: [header]
...